// packages/website/src/graphql/typings/FetchModules.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: FetchModules
// ====================================================

export interface FetchModules_fetchModules {
  __typename: "GlobalType";
  commandGroups: string[] | null;
  commandDescriptions: string[] | null;
}

export interface FetchModules {
  fetchModules: FetchModules_fetchModules;
}